requirements.txt
streamlit
pandas
numpy
scikit-learn
supabase
requests
boto3
google-auth
google-auth-oauthlib
streamlit-javascript
PyJWT
cryptography
